implementa funções do service de Projeto

+++ Framework/src/main/java/framework/service/ServiceProjeto.java
@@ -30,16 +30,36 @@ public abstract class ServiceProjeto<P extends Projeto> implements IServiceProje
 	
 	@Override
 	public void remover(P p) throws DatabaseException {
+		if (p == null) 
+			throw new IllegalArgumentException("Parâmetro fornecido nulo");
+		if (p.getCodigo() == null )
+			throw new IllegalArgumentException("Identificador de projeto nulo!");
+		
+		validarRemover(p);
 		dao.remover(p);
 	}
 	
 	@Override
 	public void atualizar(P p) throws DatabaseException {
+		if (p == null)
+			throw new IllegalArgumentException("Parâmetro fornecido nulo");
+		if (p.getCodigo() == null) 
+			throw new IllegalArgumentException("Identificador de projeto nulo!");
+		if(p.getDataInicio() == null)
+			throw new IllegalArgumentException("Data inicio nula!");
+		if(p.getNome() == null)
+			throw new IllegalArgumentException("Nome nulo!");
+		
+		validarAtulizar(p);
 		dao.atualizar(p);
 	}
 	
 	@Override
 	public List<P> consultar(P p) throws DatabaseException{
+		if (p == null)
+			throw new IllegalArgumentException("Parâmetro fornecido nulo");
+			
+		validarConsultar(p);
 		return dao.consultar(p);
 	}
 	
@@ -54,4 +74,4 @@ public abstract class ServiceProjeto<P extends Projeto> implements IServiceProje
 	protected abstract void validarAtulizar(P p);
 	protected abstract void validarConsultar(P p);
 
-}
+}
